reporter: Implement the `EvaluatedModel`
The `EvaluatedModel` contains the data from the `OrtResult` in an
evaluated format where e.g. excludes and resolutions are applied.
The purpose of this is to provide a model that can be used by external
tools without having to re-implement ORT algorithms. The main user of
the model will be the `WebAppReporter`, which currently does not support
excludes and resolutions, because the required algorithms are not
available in Javascript.

To reduce duplication in the model Jackson's `ObjectIdInfo` feature is
used, which replaces object references with generated identifiers.

`LicenseFindingCurations` are not yet implemented by this commit and
will be added later. Also the test data currently used does not test all
features of the model and needs to be improved.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@here.com>