FindingsMatcher: Discard less copyright statements
The current algorithm links copyright statements detected in files without
any detected license to all root license. At the same time copyright
statements detected in files with multiple detected licenses are discarded
if the license locations are not nearby. Let's consistently assign such
copyright statements to all root licenses in order to move a bit from
the under towards the overreporting which better matches ORTs philosophy.
As result copyright statements aren't discarded anymore in case there is
at least one root license detected.

I recall some historical reasoning behind that discarding, it said that
the likelyhood of such discarded statements being false positive
detections is signifficantly higher. Whether this argument is sufficient
`to be legally on the safe side` is a matter of opinion, my tests with a
large scan indicated that most of the discarded copyrights statements
actually are not false positives.

Note: The ultimate goal should be to (optionally) not discard any
copyright statements at all. That's a bit harder to achieve which is why
this change is limited to that low hanging fruit.

Signed-off-by: Frank Viernau <frank.viernau@here.com>